+++ packages/ingest/util.js
@@ -2,24 +2,6 @@
 
 const mime = require('mime-types');
 
-const { deprecate } = require('@cumulus/common/util');
-
-/**
- * Ensure provider path conforms to expectations.
- * Removes any/all leading forward slashes.
- *
- * @param {string} provPath - provider path
- * @returns {string} path, updated to conform if necessary.
- */
-function normalizeProviderPath(provPath) {
-  deprecate('normalizeProviderPath', 'v1.18.0', 'path.normalize');
-  if (provPath) {
-    const leadingSlashRegex = /^\/*/g;
-    return provPath.replace(leadingSlashRegex, '');
-  }
-  return '';
-}
-
 /**
  * Return mime-type based on input url or filename
  *
@@ -31,6 +13,5 @@ function lookupMimeType(key) {
 }
 
 module.exports = {
-  normalizeProviderPath,
   lookupMimeType
 };