+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### PLEASE NOTE
+
+- **CUMULUS-1762**: the `messageConsumer` for `sns` and `kinesis`-type rules now fetches
+  the collection information from the message. You should ensure that your rule's collection
+  name and version match what is in the message for these ingest messages to be processed.
+  If no matching rule is found, an error will be thrown and logged in the
+  `messageConsumer` Lambda function's log group.
+
 ### Added
 
 - **CUMULUS-1629**`
@@ -16,6 +24,19 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Adds a new copy of the API lambda `PrivateApiLambda() configured to not require auth, without an API front end
   - Adds @cumulus/api-client with functions for use by workflow lambdas to call the API when needed
 
+
+- **CUMULUS-1732**
+  - Added Python task/activity workflow and integration test (`PythonReferenceSpec`) to test `cumulus-message-adapter-python`and `cumulus-process-py` integration.
+
+- **CUMULUS-1795**
+  - Added an IAM policy on the Cumulus EC2 creation to enable SSM when the `deploy_to_ngap` flag is true
+
+### Changed
+
+- **CUMULUS-1762**
+  - the `messageConsumer` for `sns` and `kinesis`-type rules now fetches the collection
+    information from the message.
+
 ### Deprecated
 
 - **CUMULUS-1629**
@@ -23,27 +44,90 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Removed
 
-- **CUMULUS-1799**
+- **CUMULUS-1799**: Deprecated code removals
   - Removed deprecated method `@cumulus/api/models/Granule.createGranulesFromSns()`
   - Removed deprecated method `@cumulus/api/models/Granule.removeGranuleFromCmr()`
-  - Removed `@cumulus/cmrjs/cmr-utils/getGranuleId()`
-  - Removed `@cumulus/cmrjs/cmr-utils/getCmrFiles()`
+  - Removed from `@cumulus/common/aws`:
+    - `apigateway()`
+    - `buildS3Uri()`
+    - `calculateS3ObjectChecksum()`
+    - `cf()`
+    - `cloudwatch()`
+    - `cloudwatchevents()`
+    - `cloudwatchlogs()`
+    - `createAndWaitForDynamoDbTable()`
+    - `createQueue()`
+    - `deleteSQSMessage()`
+    - `describeCfStackResources()`
+    - `downloadS3File()`
+    - `downloadS3Files()`
+    - `DynamoDbSearchQueue` class
+    - `dynamodbstreams()`
+    - `ec2()`
+    - `ecs()`
+    - `fileExists()`
+    - `findResourceArn()`
+    - `fromSfnExecutionName()`
+    - `getFileBucketAndKey()`
+    - `getJsonS3Object()`
+    - `getQueueUrl()`
+    - `getObjectSize()`
+    - `getS3ObjectReadStream()`
+    - `getSecretString()`
+    - `getStateMachineArn()`
+    - `headObject()`
+    - `isThrottlingException()`
+    - `kinesis()`
+    - `lambda()`
+    - `listS3Objects()`
+    - `promiseS3Upload()`
+    - `publishSnsMessage()`
+    - `putJsonS3Object()`
+    - `receiveSQSMessages()`
+    - `s3CopyObject()`
+    - `s3GetObjectTagging()`
+    - `s3Join()`
+    - `S3ListObjectsV2Queue` class
+    - `s3TagSetToQueryString()`
+    - `s3PutObjectTagging()`
+    - `secretsManager()`
+    - `sendSQSMessage()`
+    - `sfn()`
+    - `sns()`
+    - `sqs()`
+    - `sqsQueueExists()`
+    - `toSfnExecutionName()`
+    - `uploadS3FileStream()`
+    - `uploadS3Files()`
+    - `validateS3ObjectChecksum()`
+  - Removed `@cumulus/common/CloudFormationGateway` class
+  - Removed `@cumulus/common/concurrency/Mutex` class
+  - Removed `@cumulus/common/errors`
+  - Removed `@cumulus/common/sftp`
+  - Removed `@cumulus/common/string.unicodeEscape`
+  - Removed `@cumulus/cmrjs/cmr-utils.getGranuleId()`
+  - Removed `@cumulus/cmrjs/cmr-utils.getCmrFiles()`
   - Removed `@cumulus/cmrjs/cmr/CMR` class
   - Removed `@cumulus/cmrjs/cmr/CMRSearchConceptQueue` class
-  - Removed `@cumulus/cmrjs/utils/getHost()`
-  - Removed `@cumulus/cmrjs/utils/getIp()`
-  - Removed `@cumulus/cmrjs/utils/hostId()`
-  - Removed `@cumulus/cmrjs/utils/updateToken()`
+  - Removed `@cumulus/cmrjs/utils.getHost()`
+  - Removed `@cumulus/cmrjs/utils.getIp()`
+  - Removed `@cumulus/cmrjs/utils.hostId()`
   - Removed `@cumulus/cmrjs/utils/ummVersion()`
-  - Removed `@cumulus/cmrjs/utils/validateUMMG()`
-  - Removed `@cumulus/ingest/aws/getEndpoint()`
-  - Removed `@cumulus/ingest/aws/getExecutionUrl()`
+  - Removed `@cumulus/cmrjs/utils.updateToken()`
+  - Removed `@cumulus/cmrjs/utils.validateUMMG()`
+  - Removed `@cumulus/ingest/aws.getEndpoint()`
+  - Removed `@cumulus/ingest/aws.getExecutionUrl()`
   - Removed `@cumulus/ingest/aws/invoke()`
   - Removed `@cumulus/ingest/aws/CloudWatch` class
   - Removed `@cumulus/ingest/aws/ECS` class
   - Removed `@cumulus/ingest/aws/Events` class
   - Removed `@cumulus/ingest/aws/SQS` class
   - Removed `@cumulus/ingest/aws/StepFunction` class
+  - Removed `@cumulus/ingest/util.normalizeProviderPath()`
+  - Removed `@cumulus/integration-tests/index.listCollections()`
+  - Removed `@cumulus/integration-tests/index.listProviders()`
+  - Removed `@cumulus/integration-tests/index.rulesList()`
+  - Removed `@cumulus/integration-tests/api/api.addCollectionApi()`
 
 ## [v1.20.0] 2020-03-12
 
@@ -234,6 +318,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   - Added `@cumulus/aws-client/CloudFormation.getCfStackParameterValues()` to get multiple parameter values for a Cloudformation stack
 
 ### Changed
+
+- **CUMULUS-1725**
+  - Moved the logic that updates the granule files cache Dynamo table into its
+    own Lambda function called `granuleFilesCacheUpdater`.
 
 - **CUMULUS-1736**
   - The `collections` model in the API package now determines the name of a